# api/openapi.yaml
openapi: 3.0.3

info:
  title: Kurso API
  description: |
    Публичный, админский и партнёрский API платформы Kurso.
    Этот файл — единый источник правды для контракта между `kurso-api` (Go)
    и фронтендами (`kurso-web`, `kurso-admin`, `kurso-partner`).
    На этапе 0 описаны только базовые типы и health-check.
  version: 0.1.0
  contact:
    name: Kurso
    url: https://kurso.io

servers:
  - url: http://localhost:8080
    description: Локальная разработка
  - url: https://api.kurso.io
    description: Production

tags:
  - name: system
    description: Служебные эндпоинты (health, readiness)
  - name: public
    description: Публичный API (kurso-web)

paths:
  /healthz:
    get:
      tags: [system]
      summary: Liveness probe
      description: Проверка того, что процесс жив. Не проверяет зависимости.
      operationId: getHealth
      responses:
        '200':
          description: Сервис жив
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'

  /readyz:
    get:
      tags: [system]
      summary: Readiness probe
      description: Проверка готовности принимать трафик (БД, Redis доступны).
      operationId: getReadiness
      responses:
        '200':
          description: Сервис готов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Readiness'
        '503':
          description: Сервис не готов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Readiness'

  /api/v1/currencies:
    get:
      tags: [public]
      summary: Список валют
      operationId: listCurrencies
      responses:
        '200':
          description: Список валют
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Currency'
        default:
          $ref: '#/components/responses/Error'

  /api/v1/directions:
    get:
      tags: [public]
      summary: Список направлений
      operationId: listDirections
      responses:
        '200':
          description: Список направлений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Direction'
        default:
          $ref: '#/components/responses/Error'

  /api/v1/exchangers:
    get:
      tags: [public]
      summary: Список обменников
      operationId: listExchangers
      responses:
        '200':
          description: Список обменников
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Exchanger'
        default:
          $ref: '#/components/responses/Error'

  /api/v1/exchangers/{slug}:
    get:
      tags: [public]
      summary: Карточка обменника
      operationId: getExchanger
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Обменник
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exchanger'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'

  /api/v1/rates/{direction}:
    get:
      tags: [public]
      summary: Курсы по направлению
      description: Текущие курсы обменников по направлению, лучший — первым.
      operationId: getRates
      parameters:
        - name: direction
          in: path
          required: true
          description: Слаг направления, например usdt-tinkoff
          schema:
            type: string
      responses:
        '200':
          description: Курсы по направлению
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatesResponse'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'

  /api/v1/rates/{direction}/stream:
    get:
      tags: [public]
      summary: Поток обновлений курсов (SSE)
      description: |
        Server-Sent Events. Каждое событие `rates` содержит актуальный
        массив `RateRow` по направлению. Курсы обновляются на сервере
        и пушатся сюда без опроса.
      operationId: streamRates
      parameters:
        - name: direction
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Поток событий SSE
          content:
            text/event-stream:
              schema:
                type: string

components:
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
          example: ok

    Readiness:
      type: object
      required: [status, checks]
      properties:
        status:
          type: string
          enum: [ok, degraded]
        checks:
          type: object
          additionalProperties:
            type: string
          example:
            postgres: ok
            redis: ok

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Машиночитаемый код ошибки
          example: not_found
        message:
          type: string
          description: Человекочитаемое сообщение
        details:
          type: object
          additionalProperties: true
          description: Дополнительный контекст ошибки

    Pagination:
      type: object
      required: [total, limit, offset]
      properties:
        total:
          type: integer
          format: int64
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 200
        offset:
          type: integer
          minimum: 0

    CurrencyKind:
      type: string
      description: Тип валюты
      enum: [crypto, fiat, cash]

    Currency:
      type: object
      required: [id, code, name, kind]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          description: Тикер/код валюты
          example: BTC
        name:
          type: string
          example: Bitcoin
        kind:
          $ref: '#/components/schemas/CurrencyKind'
        network:
          type: string
          nullable: true
          description: Сеть для крипты (TRC20, ERC20 и т.п.)
          example: TRC20
        iconUrl:
          type: string
          format: uri
          nullable: true

    Direction:
      type: object
      required: [id, fromCurrencyId, toCurrencyId, slug, fromCode, fromName, toCode, toName, isPopular]
      properties:
        id:
          type: string
          format: uuid
        fromCurrencyId:
          type: string
          format: uuid
        toCurrencyId:
          type: string
          format: uuid
        slug:
          type: string
          example: usdt-tinkoff
        fromCode:
          type: string
          example: USDT
        fromName:
          type: string
          example: Tether
        toCode:
          type: string
          example: TINKOFF
        toName:
          type: string
          example: Тинькофф
        isPopular:
          type: boolean

    ExchangerStatus:
      type: string
      enum: [active, paused, banned]

    Exchanger:
      type: object
      required: [id, slug, name, status, partner, isVerified, reviewsCount, directionsCount, assets, onSince]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          example: cryptobridge
        name:
          type: string
          example: CryptoBridge
        status:
          $ref: '#/components/schemas/ExchangerStatus'
        partner:
          type: boolean
          description: Есть партнёрское размещение (реферальная ссылка)
        isVerified:
          type: boolean
        reviewsCount:
          type: integer
        ratingAvg:
          type: number
          format: float
          nullable: true
          minimum: 0
          maximum: 5
        websiteUrl:
          type: string
          format: uri
          nullable: true
        logoUrl:
          type: string
          format: uri
          nullable: true
        description:
          type: string
          nullable: true
        reserveTotal:
          type: string
          nullable: true
          description: Суммарный резерв по активным курсам (decimal-строкой), null если курсов нет
        directionsCount:
          type: integer
          description: Число направлений с активным курсом
        assets:
          type: array
          items:
            type: string
          description: Торгуемые исходные валюты (коды), напр. USDT, BTC
        onSince:
          type: integer
          description: Год добавления обменника на Kurso

    RateRow:
      type: object
      description: Курс конкретного обменника по направлению (денормализовано для выдачи).
      required: [exchangerId, exchangerSlug, exchangerName, partner, reviewsCount, rate, fetchedAt]
      properties:
        exchangerId:
          type: string
          format: uuid
        exchangerSlug:
          type: string
        exchangerName:
          type: string
        partner:
          type: boolean
        ratingAvg:
          type: number
          format: float
          nullable: true
        reviewsCount:
          type: integer
        rate:
          type: string
          description: Курс decimal-строкой
          example: '81.20'
        reserve:
          type: string
          nullable: true
        minAmount:
          type: string
          nullable: true
        maxAmount:
          type: string
          nullable: true
        fetchedAt:
          type: string
          format: date-time

    RatesResponse:
      type: object
      required: [direction, rates]
      properties:
        direction:
          $ref: '#/components/schemas/Direction'
        rates:
          type: array
          items:
            $ref: '#/components/schemas/RateRow'

    Rate:
      type: object
      required: [exchangerId, directionId, rate, updatedAt]
      properties:
        exchangerId:
          type: string
          format: uuid
        directionId:
          type: string
          format: uuid
        rate:
          type: string
          description: Курс в виде decimal-строки (чтобы не терять точность)
          example: '0.0000234'
        reserve:
          type: string
          nullable: true
          description: Резерв в валюте назначения (decimal-строка)
        minAmount:
          type: string
          nullable: true
        maxAmount:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time

  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'