# kurso-api/internal/adapter/http/openapi/types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package openapi

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for CurrencyKind.
const (
	Cash   CurrencyKind = "cash"
	Crypto CurrencyKind = "crypto"
	Fiat   CurrencyKind = "fiat"
)

// Valid indicates whether the value is a known member of the CurrencyKind enum.
func (e CurrencyKind) Valid() bool {
	switch e {
	case Cash:
		return true
	case Crypto:
		return true
	case Fiat:
		return true
	default:
		return false
	}
}

// Defines values for ExchangerStatus.
const (
	Active ExchangerStatus = "active"
	Banned ExchangerStatus = "banned"
	Paused ExchangerStatus = "paused"
)

// Valid indicates whether the value is a known member of the ExchangerStatus enum.
func (e ExchangerStatus) Valid() bool {
	switch e {
	case Active:
		return true
	case Banned:
		return true
	case Paused:
		return true
	default:
		return false
	}
}

// Defines values for HealthStatus.
const (
	HealthStatusOk HealthStatus = "ok"
)

// Valid indicates whether the value is a known member of the HealthStatus enum.
func (e HealthStatus) Valid() bool {
	switch e {
	case HealthStatusOk:
		return true
	default:
		return false
	}
}

// Defines values for ReadinessStatus.
const (
	ReadinessStatusDegraded ReadinessStatus = "degraded"
	ReadinessStatusOk       ReadinessStatus = "ok"
)

// Valid indicates whether the value is a known member of the ReadinessStatus enum.
func (e ReadinessStatus) Valid() bool {
	switch e {
	case ReadinessStatusDegraded:
		return true
	case ReadinessStatusOk:
		return true
	default:
		return false
	}
}

// Currency defines model for Currency.
type Currency struct {
	// Code Тикер/код валюты
	Code    string             `json:"code"`
	IconUrl *string            `json:"iconUrl,omitempty"`
	Id      openapi_types.UUID `json:"id"`

	// Kind Тип валюты
	Kind CurrencyKind `json:"kind"`
	Name string       `json:"name"`

	// Network Сеть для крипты (TRC20, ERC20 и т.п.)
	Network *string `json:"network,omitempty"`
}

// CurrencyKind Тип валюты
type CurrencyKind string

// Direction defines model for Direction.
type Direction struct {
	FromCode       string             `json:"fromCode"`
	FromCurrencyId openapi_types.UUID `json:"fromCurrencyId"`
	FromName       string             `json:"fromName"`
	Id             openapi_types.UUID `json:"id"`
	IsPopular      bool               `json:"isPopular"`
	Slug           string             `json:"slug"`
	ToCode         string             `json:"toCode"`
	ToCurrencyId   openapi_types.UUID `json:"toCurrencyId"`
	ToName         string             `json:"toName"`
}

// Error defines model for Error.
type Error struct {
	// Code Машиночитаемый код ошибки
	Code string `json:"code"`

	// Details Дополнительный контекст ошибки
	Details *map[string]interface{} `json:"details,omitempty"`

	// Message Человекочитаемое сообщение
	Message string `json:"message"`
}

// Exchanger defines model for Exchanger.
type Exchanger struct {
	// Assets Торгуемые исходные валюты (коды), напр. USDT, BTC
	Assets      []string `json:"assets"`
	Description *string  `json:"description,omitempty"`

	// DirectionsCount Число направлений с активным курсом
	DirectionsCount int                `json:"directionsCount"`
	Id              openapi_types.UUID `json:"id"`
	IsVerified      bool               `json:"isVerified"`
	LogoUrl         *string            `json:"logoUrl,omitempty"`
	Name            string             `json:"name"`

	// OnSince Год добавления обменника на Kurso
	OnSince int `json:"onSince"`

	// Partner Есть партнёрское размещение (реферальная ссылка)
	Partner   bool     `json:"partner"`
	RatingAvg *float32 `json:"ratingAvg,omitempty"`

	// ReserveTotal Суммарный резерв по активным курсам (decimal-строкой), null если курсов нет
	ReserveTotal *string         `json:"reserveTotal,omitempty"`
	ReviewsCount int             `json:"reviewsCount"`
	Slug         string          `json:"slug"`
	Status       ExchangerStatus `json:"status"`
	WebsiteUrl   *string         `json:"websiteUrl,omitempty"`
}

// ExchangerStatus defines model for ExchangerStatus.
type ExchangerStatus string

// Health defines model for Health.
type Health struct {
	Status HealthStatus `json:"status"`
}

// HealthStatus defines model for Health.Status.
type HealthStatus string

// Pagination defines model for Pagination.
type Pagination struct {
	Limit  int   `json:"limit"`
	Offset int   `json:"offset"`
	Total  int64 `json:"total"`
}

// Rate defines model for Rate.
type Rate struct {
	DirectionId openapi_types.UUID `json:"directionId"`
	ExchangerId openapi_types.UUID `json:"exchangerId"`
	MaxAmount   *string            `json:"maxAmount,omitempty"`
	MinAmount   *string            `json:"minAmount,omitempty"`

	// Rate Курс в виде decimal-строки (чтобы не терять точность)
	Rate string `json:"rate"`

	// Reserve Резерв в валюте назначения (decimal-строка)
	Reserve   *string   `json:"reserve,omitempty"`
	UpdatedAt time.Time `json:"updatedAt"`
}

// RateRow Курс конкретного обменника по направлению (денормализовано для выдачи).
type RateRow struct {
	ExchangerId   openapi_types.UUID `json:"exchangerId"`
	ExchangerName string             `json:"exchangerName"`
	ExchangerSlug string             `json:"exchangerSlug"`
	FetchedAt     time.Time          `json:"fetchedAt"`
	MaxAmount     *string            `json:"maxAmount,omitempty"`
	MinAmount     *string            `json:"minAmount,omitempty"`
	Partner       bool               `json:"partner"`

	// Rate Курс decimal-строкой
	Rate         string   `json:"rate"`
	RatingAvg    *float32 `json:"ratingAvg,omitempty"`
	Reserve      *string  `json:"reserve,omitempty"`
	ReviewsCount int      `json:"reviewsCount"`
}

// RatesResponse defines model for RatesResponse.
type RatesResponse struct {
	Direction Direction `json:"direction"`
	Rates     []RateRow `json:"rates"`
}

// Readiness defines model for Readiness.
type Readiness struct {
	Checks map[string]string `json:"checks"`
	Status ReadinessStatus   `json:"status"`
}

// ReadinessStatus defines model for Readiness.Status.
type ReadinessStatus string